initial commit with tests working
